requirements.txt
python-dotenv
openai 
elevenlabs
assemblyai
queuelib